# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TravelCopilot API",
    "description": "Multi-Agent AI Trip Planner",
    "version": "0.1.0"
  },
  "paths": {
    "/api/trip/plan": {
      "post": {
        "tags": [
          "trip"
        ],
        "summary": "Plan Trip",
        "description": "Start a planning session and stream SSE events.",
        "operationId": "plan_trip_api_trip_plan_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/trip/{session_id}": {
      "get": {
        "tags": [
          "trip"
        ],
        "summary": "Get Itinerary",
        "description": "Return the latest itinerary for a session from the database.",
        "operationId": "get_itinerary_api_trip__session_id__get",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Itinerary Api Trip  Session Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/trip/{trip_id}/itinerary": {
      "put": {
        "tags": [
          "trip"
        ],
        "summary": "Update Itinerary",
        "description": "Persist drag-drop day reorders from the frontend.",
        "operationId": "update_itinerary_api_trip__trip_id__itinerary_put",
        "parameters": [
          {
            "name": "trip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Trip Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ItineraryUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Update Itinerary Api Trip  Trip Id  Itinerary Put"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/trip/{trip_id}/usage": {
      "get": {
        "tags": [
          "trip"
        ],
        "summary": "Get Usage",
        "description": "Return per-agent token and cost breakdown from the database.",
        "operationId": "get_usage_api_trip__trip_id__usage_get",
        "parameters": [
          {
            "name": "trip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Trip Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Usage Api Trip  Trip Id  Usage Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/trip/public/{slug}": {
      "get": {
        "tags": [
          "trip"
        ],
        "summary": "Get Public Itinerary",
        "description": "Return a public shared itinerary by slug.",
        "operationId": "get_public_itinerary_api_trip_public__slug__get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Public Itinerary Api Trip Public  Slug  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/trip/{trip_id}/pdf": {
      "post": {
        "tags": [
          "trip"
        ],
        "summary": "Generate Pdf",
        "description": "Generate a PDF of the itinerary using WeasyPrint.",
        "operationId": "generate_pdf_api_trip__trip_id__pdf_post",
        "parameters": [
          {
            "name": "trip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Trip Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/trip/{trip_id}/feedback": {
      "post": {
        "tags": [
          "trip"
        ],
        "summary": "Submit Feedback",
        "description": "Record user feedback score in Langfuse and DB.",
        "operationId": "submit_feedback_api_trip__trip_id__feedback_post",
        "parameters": [
          {
            "name": "trip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Trip Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedbackRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Submit Feedback Api Trip  Trip Id  Feedback Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/user/profile": {
      "put": {
        "tags": [
          "user"
        ],
        "summary": "Upsert Profile",
        "description": "Create or update the user profile for a session.",
        "operationId": "upsert_profile_api_user_profile_put",
        "parameters": [
          {
            "name": "X-Session-ID",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Session-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserProfile"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Upsert Profile Api User Profile Put"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "user"
        ],
        "summary": "Get Profile",
        "description": "Retrieve the user profile for the current session.",
        "operationId": "get_profile_api_user_profile_get",
        "parameters": [
          {
            "name": "X-Session-ID",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Session-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Profile Api User Profile Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "tags": [
          "ops"
        ],
        "summary": "Health",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object",
                  "title": "Response Health Health Get"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "BudgetTier": {
        "type": "string",
        "enum": [
          "budget",
          "mid",
          "luxury"
        ],
        "title": "BudgetTier"
      },
      "FeedbackRequest": {
        "properties": {
          "rating": {
            "type": "integer",
            "title": "Rating",
            "description": "1 = positive, -1 = negative"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          }
        },
        "type": "object",
        "required": [
          "rating"
        ],
        "title": "FeedbackRequest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HotelStyle": {
        "type": "string",
        "enum": [
          "hostel",
          "budget",
          "boutique",
          "business",
          "luxury"
        ],
        "title": "HotelStyle"
      },
      "ItineraryUpdateRequest": {
        "properties": {
          "trip_days": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Trip Days"
          }
        },
        "type": "object",
        "title": "ItineraryUpdateRequest"
      },
      "PlanRequest": {
        "properties": {
          "query": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 3,
            "title": "Query"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "destination": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Destination"
          }
        },
        "type": "object",
        "required": [
          "query"
        ],
        "title": "PlanRequest"
      },
      "UserProfile": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "nationality": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nationality"
          },
          "passport_country": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Passport Country"
          },
          "preferred_currency": {
            "type": "string",
            "title": "Preferred Currency",
            "default": "INR"
          },
          "dietary_restrictions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Dietary Restrictions"
          },
          "accessibility_needs": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Accessibility Needs"
          },
          "interests": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Interests"
          },
          "preferred_airlines": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Preferred Airlines"
          },
          "preferred_hotel_chains": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Preferred Hotel Chains"
          },
          "hotel_style": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/HotelStyle"
              },
              {
                "type": "null"
              }
            ]
          },
          "travel_style": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Travel Style"
          },
          "fitness_level": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fitness Level"
          },
          "altitude_experience": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Altitude Experience"
          }
        },
        "type": "object",
        "required": [
          "user_id"
        ],
        "title": "UserProfile"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}